kinhdich/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
torch
numpy
pydantic
psycopg2-binary 
pydantic-settings>=2.0
google-genai
openai
